Make optimizations to getSession operations
In frontend/app/u/home/page.jsx:
Instead of calling getSession every time the page number is updated,
Using useEffect, and useState, I am storing the session in a state, and
when the page is updated, again using useEffect to fetch the
needed data.

In frontend/app/u/home/new/page.jsx:
Instead of calling getSession inside of the onClick handler, I am
storing both the session, and the onClick handler in a state, and using
useEffect, when the session is retrieved, the onClick handler gets
redefined to the actual function.